[TS Migration] PR E: Convert entry points, remove allowJs bridge
- App.js → App.tsx: clean JSX-only conversion
- index.js → index.tsx: non-null assertion on getElementById('root')
  since #root is guaranteed present in index.html
- tsconfig.json: remove allowJs and checkJs — no .js files remain in src/
- vite.config.ts: remove load-js-files-as-jsx plugin — no longer needed

Migration complete: all 11 source files now TypeScript.

Verification: npm run build ✓  tsc --noEmit ✓  npm test ✓
