--- src/index.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-import { HashRouter } from 'react-router-dom';
-import './index.css';
-import App from './App';
-
-const root = createRoot(document.getElementById('root'));
-root.render(
-  <HashRouter>
-    <React.StrictMode>
-      <App />
-    </React.StrictMode>
-  </HashRouter>
-);
