+++ src/index.tsx
@@ -1,0 +1,15 @@
+import React from 'react'
+import { createRoot } from 'react-dom/client'
+import { HashRouter } from 'react-router-dom'
+import './index.css'
+import App from './App'
+
+// Non-null assertion: #root is guaranteed to exist in index.html
+const root = createRoot(document.getElementById('root')!)
+root.render(
+  <HashRouter>
+    <React.StrictMode>
+      <App />
+    </React.StrictMode>
+  </HashRouter>
+)
