+++ vite.config.ts
@@ -1,6 +1,6 @@
 // TODO: add ESLint + @typescript-eslint once JS→TS migration is complete (issue #17)
 /// <reference types="vitest" />
-import { defineConfig, transformWithEsbuild } from 'vite'
+import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 
 export default defineConfig({
@@ -10,21 +10,7 @@ export default defineConfig({
     setupFiles: './src/setupTests.ts',
     passWithNoTests: true,
   },
-  optimizeDeps: {
-    esbuildOptions: {
-      loader: {
-        '.js': 'jsx',
-      },
-    },
-  },
   plugins: [
-    {
-      name: 'load-js-files-as-jsx',
-      async transform(code, id) {
-        if (!id.match(/src\/.*\.js$/)) return null
-        return transformWithEsbuild(code, id, { loader: 'jsx' })
-      },
-    },
     react(),
   ],
   base: '/portfolio-pages/',
